[Feature] Add Input Manager
