--- Assets/Code/Scripts/MouseInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ee814d4727ed3242b9d3af57fa77d7a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
